+++ CHANGELOG.md
@@ -75,3 +75,4 @@
 - MVP-392 - hook up selected industries to be deleted
 - no-ticket make food default when creating export plan
 - MVP-381 - Wagtail media and video block
+- MVP-427 - Persist products upon signup